1 folder that contains all the nodes with no subfolders
