+++ .vscode/arduino.json
@@ -1,0 +1,5 @@
+{
+    "sketch": "Nodes/Nodes.ino",
+    "configuration": "cache=on,speed=120,opt=small,maxqspi=50,usbstack=arduino,debug=off",
+    "board": "adafruit:samd:adafruit_feather_m4_can"
+}
