--- Nodes/Front_Node/src/can_conv~HEAD
@@ -1,1 +1,0 @@
-../../can_conv
