+++ Nodes/Screen_Node.h
@@ -1,6 +1,6 @@
 #include <CAN.h>
 #include <millisDelay.h>
-#include "src/can_conv/sbm_network_definition.h"
+#include "can_conv/sbm_network_definition.h"
 #define SERIAL_DEBUG_PRINTS_ON 1
 /*******************************************************************************/
