+++ Nodes/Vehicle_Dyamics_Node.h
@@ -4,7 +4,7 @@
 #include <millisDelay.h>
 #include <Adafruit_LSM6DSOX.h>
 #include <Adafruit_LIS3MDL.h>
-#include "src/can_conv/sbm_network_definition.h"
+#include "can_conv/sbm_network_definition.h"
 #define SERIAL_DEBUG_PRINTS_ON 1
 /*******************************************************************************/
